исправление. Не выбранные элементы при удалении

+++ Учет цистерн/Forms/Продукты/ProductForm.cs
@@ -95,22 +95,29 @@ namespace Учет_цистерн
 
         private void button3_Click_Delete_Product(object sender, EventArgs e)
         {
-            if (MessageBox.Show("Вы действительно хотите удалить эту запись?", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            if(dataGridView1.SelectedRows.Count > 0)
             {
-                try
+                if (MessageBox.Show("Вы действительно хотите удалить эту запись?", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                 {
-                    string DeleteCurrentProduct = "delete from d__Product where ID = " + SelectItemRow;
-                    DbConnection.DBConnect(DeleteCurrentProduct);
-                    MessageBox.Show("Продукт удалён!", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    try
+                    {
+                        string DeleteCurrentProduct = "delete from d__Product where ID = " + SelectItemRow;
+                        DbConnection.DBConnect(DeleteCurrentProduct);
+                        MessageBox.Show("Продукт удалён!", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    }
+                    catch (SqlException ex)
+                    {
+                        MessageBox.Show(ex.Message, "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    catch (Exception exp)
+                    {
+                        MessageBox.Show(exp.Message, "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
                 }
-                catch (SqlException ex)
-                {
-                    MessageBox.Show(ex.Message, "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
-                catch (Exception exp)
-                {
-                    MessageBox.Show(exp.Message, "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
+            }
+            else
+            {
+                MessageBox.Show("Для удаления записи, необходимо выбрать строку!", "", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
         }
